made search page, added names to the ids on the dashboard, and refactored code + bug fixing

+++ dashboard/transaction.php
@@ -1,4 +1,12 @@
 <?php
+
+    //check if administrator is logged in
+    session_start();
+    if (!isset($_SESSION["loggedin"])) {
+        header('Location: dashboard/login.php');
+        die();
+    }
+
     include_once("includes/dbh.inc.php");
 
     if (isset($_GET["formSearch"]) && isset($_GET["type"]) && isset($_GET["query"])) {
@@ -6,12 +14,12 @@
         $type = $_GET["type"];
         $query = $_GET["query"];
 
-        $sql = "SELECT * FROM transaction LIMIT 100";
+        $sql = "SELECT * FROM transaction";
 
         if ($query != "") {
-            $mappings = array("transaction ID"=>"id", "Full Name"=>"name", "Email"=>"email", "Address"=>"address");
+            $mappings = array("Transaction ID"=>"id", "Client ID"=>"clientId", "Healer ID"=>"healerID", "Amount"=>"amount", "Payment Method"=>"paymentMethod");
             $type = $mappings[$type];
-            $sql = "SELECT * FROM transaction WHERE $type LIKE '%$query%' LIMIT 100";
+            $sql = "SELECT * FROM transaction WHERE $type LIKE '%$query%'";
         }
 
         $result = $conn->query($sql);
@@ -20,46 +28,56 @@
 
             $rows = "";
             $count = 0;
-            $enabled = "1";
+
+            $mapClientIdsToNames = array();
+            $mapHealerIdsToNames = array();
 
             while ($row = $result->fetch_assoc()) {
-                $row["enabled"] == "1" ? $enabled = "Enabled" : $enabled = "Disabled";
+
+                $clientId = $row["clientId"];
+
+                if (!array_key_exists($clientId, $mapClientIdsToNames)) {
+                
+                    $sql2 = "SELECT name FROM client WHERE id=$clientId";
+                    $result2 = $conn->query($sql2);
+                    if ($result2->num_rows > 0) {
+                        $row2 = $result2->fetch_assoc();
+                        $mapClientIdsToNames[$clientId] = $row2["name"];
+                    }
+
+                }
+
+                $healerId = $row["healerId"];
+
+                if (!array_key_exists($healerId, $mapHealerIdsToNames)) {
+
+                    $sql3 = "SELECT name FROM healer WHERE id=$healerId";
+                    $result3 = $conn->query($sql3);
+                    if ($result3->num_rows > 0) {
+                        $row3 = $result3->fetch_assoc();
+                        $mapHealerIdsToNames[$healerId] = $row3["name"];
+                    }
+
+                }
+
                 $count++;
-                $rows .= "<tr><td><input type='checkbox' id='checkbox$count'></td><td>" . $row["id"] . "</td><td>" . $row["name"]. "</td><td>" . $row["email"] . "</td><td>" . $row["password"] . "</td><td>" . $row["address"] . "</td><td>" . $enabled . "</td></tr>";
+                $rows .= "<tr><td><input type='checkbox' id='checkbox$count'></td><td>" . $row["id"] . "</td><td>" . $clientId . " (" . $mapClientIdsToNames[$clientId] . ")</td><td>" . $healerId . " (" . $mapHealerIdsToNames[$healerId] . ")</td><td>$" . $row["amount"] . "</td><td>" . $row["paymentMethod"] . "</td></tr>";
             }
 
         } else {
-            $rows = "<tr><td colspan='7'>No results</td></tr>";
+            $rows = "<tr><td colspan='6'>No results</td></tr>";
         }
 
         $conn->close();
 
-    } else if ((isset($_GET["formEnable"]) || isset($_GET["formDisable"]) || isset($_GET["formRemove"])) && isset($_GET["hidden"]) && $_GET["hidden"] != "") {
+    } else if (isset($_GET["formRemove"]) && isset($_GET["hidden"]) && $_GET["hidden"] != "") {
 
         $ids = $_GET["hidden"];
 
-        if (isset($_GET["formEnable"]) || isset($_GET["formDisable"])) {
+        $sql = "DELETE FROM transaction WHERE id IN($ids)";
 
-            $sql = "";
-
-            if (isset($_GET["formEnable"])) {
-                $sql = "UPDATE transaction SET enabled = TRUE WHERE id IN($ids)";
-            } else if (isset($_GET["formDisable"])) {
-                $sql = "UPDATE transaction SET enabled = FALSE WHERE id IN($ids)";
-            }
-
-            if ($conn->query($sql) === FALSE) {
-                echo "Error modifying record: " . $conn->error;
-            }
-        
-        } else if (isset($_GET["formRemove"])) {
-
-            $sql = "DELETE FROM transaction WHERE id IN($ids)";
-
-            if ($conn->query($sql) === FALSE) {
-                echo "Error deleting record: " . $conn->error;
-            }
-            
+        if ($conn->query($sql) === FALSE) {
+            echo "Error deleting record: " . $conn->error;
         }
 
         $conn->close();
@@ -77,7 +95,7 @@
         <h1>Administrator Dashboard</h1>
         <nav>
             <ul>
-                <li><a href="transaction.php">Client</a></li>
+                <li><a href="client.php">Client</a></li>
                 <li><a href="healer.php">Healer</a></li>
                 <li><a href="appointment.php">Appointment</a></li>
                 <li id="selected-tab"><a href="transaction.php">Transaction</a></li>
@@ -87,50 +105,37 @@
         </nav>
         <div id="container">
             <div id="top-bar">
-                    <div id="top-bar-query">
-                        <form action="transaction.php" method="get">
-                                <select name="type">
-                                    <option name="id" <?php if (isset($type) && $type == "id") { echo("selected"); } ?>>Client ID</option>
-                                    <option name="name" <?php if (isset($type) && $type == "name") { echo("selected"); } ?>>Full Name</option>
-                                    <option name="email" <?php if (isset($type) && $type == "email") { echo("selected"); } ?>>Email</option>
-                                    <option name="address" <?php if (isset($type) && $type == "address") { echo("selected"); } ?>>Address</option>
-                                </select>
-                                <input type="text" name="query" value="<?php if (isset($query)) { echo($query); }?>">
-                                <input id="btn-search" type="submit" name="formSearch" value="Search">
-                                <select id="select-visibility-filter">
-                                <option id="show-all">Show All</option>
-                                <option name="show-enabled">Enabled</option>
-                                <option name="show-disabled">Disabled</option>
-                            </select>
-                        </form>
-                    </div>
-                    <div id="top-bar-toggles">
-                        <button id="btn-clear">Clear</button>
-                        <form id="formEnable" action="transaction.php" method="get">
-                            <input id="btn-enable" type="submit" name="formEnable" value="Enable">
-                            <input id="hiddenEnable" type="hidden" name="hidden">
-                        </form>
-                        <form id="formDisable" action="transaction.php" method="get">
-                            <input id="btn-disable" type="submit" name="formDisable" value="Disable">
-                            <input id="hiddenDisable" type="hidden" name="hidden">
-                        </form>
-                        <form  id="formRemove" action="transaction.php" method="get">
-                            <input id="btn-remove" type="submit" name="formRemove" value="Remove">
-                            <input id="hiddenRemove" type="hidden" name="hidden">
-                        </form>
-                    </div>
+                <div id="top-bar-query">
+                    <form action="transaction.php" method="get">
+                        <select name="type">
+                            <option name="id" <?php if (isset($type) && $type == "id") { echo("selected"); } ?>>Transaction ID</option>
+                            <option name="clientId" <?php if (isset($type) && $type == "clientId") { echo("selected"); } ?>>Client ID</option>
+                            <option name="healerId" <?php if (isset($type) && $type == "healerId") { echo("selected"); } ?>>Healer ID</option>
+                            <option name="amount" <?php if (isset($type) && $type == "amount") { echo("selected"); } ?>>Amount</option>
+                            <option name="paymentMethod" <?php if (isset($type) && $type == "paymentMethod") { echo("selected"); } ?>>Payment Method</option>
+                        </select>
+                        <input type="text" name="query" value="<?php if (isset($query)) { echo($query); }?>">
+                        <input id="btn-search" type="submit" name="formSearch" value="Search">
+                    </form>
+                </div>
+                <div id="top-bar-toggles">
+                    <button id="btn-clear">Clear</button>
+                    <form  id="formRemove" action="transaction.php" method="get">
+                        <input id="btn-remove" type="submit" name="formRemove" value="Remove">
+                        <input id="hiddenRemove" type="hidden" name="hidden">
+                    </form>
+                </div>
             </div>
             <div class="empty"></div>
             <table id="table-results">
                 <thead>
                     <tr>
                         <th></th>
+                        <th>Transaction ID</th>
                         <th>Client ID</th>
-                        <th>Full Name</th>
-                        <th>Email</th>
-                        <th>Password (Encrypted)</th>
-                        <th>Address</th>
-                        <th>Status</th>
+                        <th>Healer ID</th>
+                        <th>Amount</th>
+                        <th>Payment Method</th>
                     </tr>
                 </thead>
                 <tbody>
@@ -141,11 +146,7 @@
                     ?>
                 </tbody>
             </table>
-            <div id="footer">
-                <a class="page-change" href="#">Previous</a>
-                <a class="page-change" href="#">Next</a>
-            </div>
         </div>
     </body>
-    <script src="scripts/scripts.js"></script>
+    <script src="scripts/script_app.js"></script>
 </html>
